Enable Perfect Timing™ mode
The timing used for effects is now calculated exactly, rather than using
the time of the first frame after the effect should have occurred.

The main difference is that the end time of animations will now be
calculated correctly if played on a slow computer where the frame
callback was called partway through when the effect should have started.