Allow configuring hues-effects without editing the source.
Options can be passed to the setup function, which will override the
defaults. In addition, the HuesEffect object now has getters/setters for
all of the user configuration options.

Changing configuration that requires a shader recompile (like changing
blendMode) is handled transparently - the shaders will be recompiled on
the next frame render.