Add an 8-sample blur.
WebGL basically guarantees that 8 varyings will work, so this means I
don't need a fallback non-blur any more.

Require at least 10 varyings for the 9 sample blur, because ANGLE on
D3D9 seems to lie about the number of varyings it supports (claims 9,
but shaders with >8 fail to compile).

Fix sign of one of the samples on the 27 sample blur.